+++ includes/AjaxOps.php
@@ -18,7 +18,8 @@ class AjaxOps {
 	public function pprh_update_hints() {
 		if ( isset( $_POST['pprh_data'] ) && wp_doing_ajax() ) {
 			check_ajax_referer( 'pprh_table_nonce', 'nonce' );
-			$this->init( $_POST['pprh_data'] );
+			$pprh_data = json_decode( wp_unslash( $_POST['pprh_data'] ), true );
+			$this->init( $pprh_data );
 			return true;
 		}
 
@@ -28,19 +29,16 @@ class AjaxOps {
 	}
 
 	public function init( $pprh_data ) {
-		$data = json_decode( wp_unslash( $pprh_data ), true );
+		if ( is_array( $pprh_data ) ) {
+			$db_result = $this->handle_action( $pprh_data );
 
-		if ( is_array( $data ) ) {
-			$action = $data['action'];
-			$result = $this->handle_action( $data, $action );
-
-			if ( is_object( $result ) ) {
+			if ( is_object( $db_result ) ) {
 				$on_pprh_admin = Utils::on_pprh_admin();
 				$display_hints = new DisplayHints( $on_pprh_admin );
-				$json = $display_hints->ajax_response( $result );
+				$json = $display_hints->ajax_response( $db_result );
 
 				if ( $this->testing ) {
-					return $json;
+					return $db_result;
 				}
 
 				wp_die( $json );
@@ -49,17 +47,9 @@ class AjaxOps {
 	}
 
 
-	public function handle_action( $data, $action ) {
-		$dao = new DAO();
-		$concat_ids = Utils::array_into_csv( $data['hint_ids'] );
-
-		if ( preg_match( '/create|update/', $action ) ) {
-			$result = $this->create_update_hint( $data, $action );
-		} elseif ( preg_match( '/enable|disable/', $action ) ) {
-			$result = $dao->bulk_update( $concat_ids, $action );
-		} elseif ( 'delete' === $action ) {
-			$result = $dao->delete_hint( $concat_ids );
-		}
+	private function handle_action( $data ) {
+		$dao_ctrl = new DAOController();
+		$result = $dao_ctrl->hint_controller( $data );
 
 //		TODO
 //		elseif ( 'reset_single_post_preconnects' === $action ) {
@@ -72,22 +62,5 @@ class AjaxOps {
 		// TODO: if error, return generic error msg if no error from db is there.
 		return ( is_object( $result ) ? $result : false );
 	}
-
-	protected function create_update_hint( $data, $action ) {
-		$dao = new DAO();
-		$create_hints = new CreateHints();
-		$pprh_hint = $create_hints->new_hint_controller( $data );
-
-		if ( ! is_array( $pprh_hint ) ) {
-			$response = $pprh_hint;
-		} elseif ( 'create' === $action ) {
-			$response = $dao->insert_hint( $pprh_hint );
-		} else {
-			$response = $dao->update_hint( $pprh_hint, $data['hint_ids'] );
-		}
-
-		return $response;
-	}
-
 
 }