+++ includes/DAO.php
@@ -8,12 +8,24 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 class DAO {
 
+	public function code_action_arr( $code ) {
+		$actions = array(
+			0 => array( 'create', 'created' ),
+			1 => array( 'update', 'updated' ),
+			2 => array( 'delete', 'deleted' ),
+			3 => array( 'enable', 'enabled' ),
+			4 => array( 'disable', 'disabled' )
+		);
+
+		return $actions[$code];
+	}
+
 	// db results
-	public function create_db_result( $result, $hint_id, $last_error, $action = '', $new_hint = null ) {
+	public function create_db_result( $result, $hint_id, $last_error, $action_code = '', $new_hint = null ) {
 		return (object) array(
 			'new_hint'  => $new_hint,
 			'db_result' => array(
-				'msg'        => self::create_msg( $result, $last_error, $action ),
+				'msg'        => $this->create_msg( $result, $last_error, $action_code ),
 				'status'     => ( $result ) ? 'success' : 'error',
 				'hint_id'    => $hint_id,
 				'success'    => $result,
@@ -22,18 +34,21 @@ class DAO {
 		);
 	}
 
-	public static function create_msg( $result, $last_error, $action )  {
+	public function create_msg( $result, $last_error, $action_code )  {
+		$actions = $this->code_action_arr( $action_code );
+
 		if ( $result ) {
-			$action .= 'd';
-			$msg = "Resource hint $action successfully.";
+			$msg = "Resource hint $actions[1] successfully.";
 		} elseif ( '' !== $last_error ) {
 			$msg = $last_error;
 		} else {
-			$msg = "Failed to $action hint.";
+			$msg = "Failed to $actions[0] hint.";
 		}
 
 		return $msg;
 	}
+
+
 
 
 	public function insert_hint( $new_hint ) {
@@ -62,7 +77,7 @@ class DAO {
 			$args['types']
 		);
 
-		return $this->create_db_result( $wpdb->result, $wpdb->insert_id, $wpdb->last_error, 'create', $new_hint );
+		return $this->create_db_result( $wpdb->result, $wpdb->insert_id, $wpdb->last_error, 0, $new_hint );
 	}
 
 
@@ -89,19 +104,7 @@ class DAO {
 			array( '%d' )
 		);
 
-		return $this->create_db_result( $wpdb->result, $wpdb->insert_id, $wpdb->last_error, 'update', $new_hint );
-	}
-
-	public function bulk_update( $hint_ids, $action ) {
-		global $wpdb;
-		$table = PPRH_DB_TABLE;
-
-		$wpdb->query( $wpdb->prepare(
-			"UPDATE $table SET status = %s WHERE id IN ($hint_ids)",
-			$action . 'd'
-		) );
-
-		return $this->create_db_result( $wpdb->result, $wpdb->insert_id, $wpdb->last_error, $action, null );
+		return $this->create_db_result( $wpdb->result, $wpdb->insert_id, $wpdb->last_error, 1, $new_hint );
 	}
 
 	public function delete_hint( $hint_ids ) {
@@ -111,12 +114,27 @@ class DAO {
 
 		if ( $hint_id_exists > 0 ) {
 			$wpdb->query( "DELETE FROM $table WHERE id IN ($hint_ids)" );
-			return $this->create_db_result( $wpdb->result, $wpdb->insert_id, $wpdb->last_error, 'delete', null );
-		} else {
-			return $this->create_db_result( false, null, 'No hint IDs to delete.', 'delete', null );
+			return $this->create_db_result( $wpdb->result, $wpdb->insert_id, $wpdb->last_error, 2, null );
+		}
+
+		else {
+			return $this->create_db_result( false, null, 'No hint IDs to delete.', 2, null );
 		}
 	}
 
+	public function bulk_update( $hint_ids, $code ) {
+		global $wpdb;
+		$table = PPRH_DB_TABLE;
+
+		$action = ( 3 === $code ) ? 'enabled' : 'disabled';
+
+		$wpdb->query( $wpdb->prepare(
+			"UPDATE $table SET status = %s WHERE id IN ($hint_ids)",
+			$action . 'd'
+		) );
+
+		return $this->create_db_result( $wpdb->result, $wpdb->insert_id, $wpdb->last_error, $code, null );
+	}
 
 
 	public function get_all_hints( $query_code = null ) {
@@ -134,6 +152,22 @@ class DAO {
 		}
 
 		return $results;
+	}
+
+	private function parse_query_code( $sql, $query_code ) {
+		$query = array( 'sql' => $sql );
+
+		if ( 1 === $query_code ) {
+			$query['sql'] .= ' WHERE status = %s';
+			$query['args'] = array( 'enabled' );
+		} elseif ( 2 === $query_code ) {
+			$query['sql'] .= ' ORDER BY ' . esc_sql( $_REQUEST['orderby'] );
+			$query['sql'] .= ! empty( $_REQUEST['order'] ) ? ' ' . esc_sql( $_REQUEST['order'] ) : ' ASC';
+		} elseif ( 3 === $query_code ) {
+			$query['sql'] .= ' ORDER BY url ASC';
+		}
+
+		return $query;
 	}
 
 	public function parse_query_code( $sql, $query_code ) {