+++ pre-party-browser-hints.php
@@ -3,7 +3,7 @@
  * Plugin Name:       Pre* Party Resource Hints
  * Plugin URI:        https://wordpress.org/plugins/pre-party-browser-hints/
  * Description:       Take advantage of the browser resource hints DNS-Prefetch, Prerender, Preconnect, Prefetch, and Preload to improve page load time.
- * Version:           1.7.5
+ * Version:           1.7.5.2
  * Requires at least: 4.4
  * Requires PHP:      5.6.30
  * Author:            Sam Perrow
@@ -13,7 +13,7 @@
  * Text Domain:       pprh
  * Domain Path:       /languages
  *
- * last edited March 18, 2021
+ * last edited March 23, 2021
  *
  * Copyright 2016  Sam Perrow  (email : sam.perrow399@gmail.com)
  *
@@ -32,8 +32,6 @@ register_activation_hook( __FILE__, array( $pprh_load, 'activate_plugin' ) );
 add_action( 'wpmu_new_blog', array( $pprh_load, 'activate_plugin' ) );
 
 class Pre_Party_Browser_Hints {
-
-    public $all_hints = array();
 
 	public function __construct() {
 	    add_action( 'init', array( $this, 'load_plugin' ) );
@@ -92,12 +90,12 @@ class Pre_Party_Browser_Hints {
 
 	public function load_common_files() {
 		include_once 'includes/Utils.php';
-		include_once 'includes/DAO.php';
+		include_once 'includes/DAOController.php';
 		include_once 'includes/CreateHints.php';
 	}
 
 	public function check_to_upgrade() {
-		$desired_version = '1.7.5';
+		$desired_version = '1.7.5.1';
 		$current_version = get_option( 'pprh_version' );
 
 		if ( $desired_version !== $current_version ) {
@@ -108,7 +106,7 @@ class Pre_Party_Browser_Hints {
 	public function activate_plugin() {
 		if ( ! class_exists( \PPRH\Utils::class ) ) {
 			include_once 'includes/Utils.php';
-			include_once 'includes/DAO.php';
+			include_once 'includes/DAOController.php';
 		}
 
 		$this->create_constants();