+++ tests/AjaxOpsTest.php
@@ -3,16 +3,26 @@
 declare(strict_types=1);
 use PHPUnit\Framework\TestCase;
 
-if (!defined('ABSPATH')) {
+if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
 final class AjaxOpsTest extends TestCase {
+	
+	public function test_start() {
+		if ( ! WP_ADMIN ) return;
 
-	public function test_pprh_update_hints(): void {
-		if ( ! WP_ADMIN || ! wp_doing_ajax() ) return;
+		$this->eval_pprh_update_hints();
+		$this->eval_init_1();
+		$this->eval_init_2();
+		$this->eval_init_3();
+		$this->eval_init_4();
+		$this->eval_init_5();
 
-		$expected = true;
+	}
+
+	public function eval_pprh_update_hints(): void {
+		$expected = wp_doing_ajax();
 		$ajax_ops = new \PPRH\AjaxOps();
 		$expected_nonce = TestUtils::create_nonce('pprh_table_nonce');
 		$_POST['pprh_data'] = '{"url":"tester","hint_type":"dns-prefetch","action":"create","hint_ids":null"}';
@@ -24,14 +34,71 @@ final class AjaxOpsTest extends TestCase {
 
 	// also need to verify update, delete, enable, disable, bulk operations work properly.
 
-//	public function test_init():void {
-//		$dao = new \PPRH\DAO();
-//		$response = json_decode( $json, true );
-//		$db_result = $response['result']['db_result'];
-//		$result = $db_result['success'];
-//		$hint_id = $db_result['hint_id'];
-//	}
+	public function eval_init_1():void {
+		$ajax_ops = new \PPRH\AjaxOps();
+		$dao_ctrl = new \PPRH\DAOController();
 
+		$test_data_1 = TestUtils::create_hint_array( 'https://testAjaxOps1.com', 'dns-prefetch' );
+		$test_data_1['op_code'] = 0;
+
+		$actual = $ajax_ops->init( $test_data_1 );
+		$expected = $dao_ctrl->create_db_result( true, '', '', $test_data_1['op_code'], $actual->new_hint );
+		$this->assertEquals($expected, $actual);
+	}
+
+	public function eval_init_2():void {
+		$ajax_ops = new \PPRH\AjaxOps();
+		$dao_ctrl = new \PPRH\DAOController();
+
+		$test_data_1 = TestUtils::create_hint_array( 'https://testAjaxOps2.com', 'dns-prefetch' );
+		$test_data_1['op_code'] = 1;
+
+		$actual = $ajax_ops->init( $test_data_1 );
+		$expected = $dao_ctrl->create_db_result( true, '', '', $test_data_1['op_code'], $actual->new_hint );
+		$this->assertEquals($expected, $actual);
+	}
+
+	public function eval_init_3():void {
+		$ajax_ops = new \PPRH\AjaxOps();
+		$dao_ctrl = new \PPRH\DAOController();
+
+		$test_data_1 = array(
+			'op_code'  => 2,
+			'hint_ids' => ''
+		);
+
+		$actual = $ajax_ops->init( $test_data_1 );
+		$expected = $dao_ctrl->create_db_result( true, '', '', $test_data_1['op_code'], null );
+		$this->assertEquals($expected, $actual);
+	}
+
+	public function eval_init_4():void {
+		$ajax_ops = new \PPRH\AjaxOps();
+		$dao_ctrl = new \PPRH\DAOController();
+
+		$test_data_1 = array(
+			'op_code'  => 3,
+			'hint_ids' => '100'
+		);
+
+		$actual = $ajax_ops->init( $test_data_1 );
+		$expected = $dao_ctrl->create_db_result( true, $test_data_1['hint_ids'], '', $test_data_1['op_code'], null );
+		$this->assertEquals($expected, $actual);
+	}
+
+	public function eval_init_5():void {
+		$ajax_ops = new \PPRH\AjaxOps();
+		$dao_ctrl = new \PPRH\DAOController();
+
+		$test_data_1 = array(
+			'op_code'  => 4,
+			'hint_ids' => '100'
+		);
+
+		$actual = $ajax_ops->init( $test_data_1 );
+		$expected = $dao_ctrl->create_db_result( true, $test_data_1['hint_ids'], '', $test_data_1['op_code'], null );
+		$this->assertEquals($expected, $actual);
+	}
 
 }